Simplify
- Move ThreadSafeStopWatch back to where it was initially.
- Perform logging after queries inside EsqlResponseListener again.
